requirements.txt
langgraph
arize-phoenix==12.14.2
psycopg2-binary
python-dotenv
langchain_google_genai
